relay: Implement a circuit cell queue maximum size
This commit introduces the consensus parameter "circ_max_cell_queue_size"
which controls the maximum number of cells a circuit queue should have.

The default value is currently 8000 + 1000 cells which is the maximum circuit
window value (1000) times the maximum number of hops on a circuit (8). We add
an extra circuit window (1000) as a marge for errors.

Related to this is #9072.

Fixes #25226

Signed-off-by: David Goulet <dgoulet@torproject.org>