Refactor the descriptor decryption/decoding
This commit refactors the existing decryption code to make it compatible
with a new logic when the client authorization is enabled.

As you can see it the changed code I made hs_desc_decode_descriptor to
receive one more parameter which is a private key of the client.

This parameter can be NULL because sometimes the client auth is disabled
or we cannot find an appropriate a private key of the client.