hs-v3: Ignore the revision counter if missing
When decoding the plaintext of a descriptor, we no long make the revision
counter mandatory.

It is also ignored by the client cache. Directory now use a replaycache
implemented in the previous commit.

Service still put it in their descriptor for backward compatibility for
clients' version 032 and 033. And because it is used in the descriptor
encryption key construction, there is really no easy way out of it for now
forcing us to keep it.

Part of #25552

Signed-off-by: David Goulet <dgoulet@torproject.org>