Load client secret key into the daemon
When the user put ClientOnionAuthDir in torrc, the daemon should load
all private keys in that directory. If some file in the directory fails
to be loaded, it just skip that file and go to the next file. But if
there is one file that has the same onion address with another file, it
will stop and fail immediately.