hs: Limit the amount of relayed INTRODUCE2
This commit introduces hs_dos.c that has the purpose of having the anti-DoS
code for onion services.

At this commit, it only has one which is a function that decides if an
INTRODUCE2 can be sent on the given introduction service circuit (S<->IP)
considering the inflight cell limitation (package window).

Basic defenses on #15516.

Signed-off-by: David Goulet <dgoulet@torproject.org>