tracing: Add LTTng UST support for cell events
This adds LTTng userspace tracing support into Tor along with a cell event
tracing subsystem.

In order to trace cell timings within the tor process, we needed a system to
track cells position in the outbuf and add a cell ID.

All this is only built with this configure option:

  --event-tracing-cells
  --event-tracing-lttng

They aren't enabled by default. If no tracers is enabled, the cell event
tracepoint are basically NOP.

Signed-off-by: David Goulet <dgoulet@torproject.org>